requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.5
djangorestframework==3.9.1
idna==2.8
more-itertools==5.0.0
mysql==0.0.2
mysqlclient==1.4.1
pluggy==0.8.1
py==1.7.0
Pygments==2.3.1
pytest==4.2.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
